:sparkles: feat: 수익률 반올림 및 출력 기능 구현
메서드
- roundsYield()
- printYield()

구현내용
- 소수점 둘째자리에서 반올림하여 반환
- Math.round 메서드를 사용

+++ src/main/java/lotto/entity/User.java
@@ -8,7 +8,7 @@ public class User {
     private List<Integer> userNumber;
     private int bonusNumber;
     private int lottoAmount;
-    private Long yield;
+    private double yield;
     private HashMap<Integer, Integer> winning = new HashMap<>(){
         {
             put(5, 0);
@@ -54,11 +54,11 @@ public class User {
         this.bonusNumber = bonusNumber;
     }
 
-    public Long getYield() {
+    public double getYield() {
         return yield;
     }
 
-    public void setYield(Long yield) {
+    public void setYield(double yield) {
         this.yield = yield;
     }
